// src/types.ts
export type Maybe<T> = T | null;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Upload: any;
};

export enum CacheControlScope {
  Public = 'PUBLIC',
  Private = 'PRIVATE'
}

export type Continent = {
   __typename?: 'Continent';
  code?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  countries?: Maybe<Array<Maybe<Country>>>;
};

export type Country = {
   __typename?: 'Country';
  code?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  native?: Maybe<Scalars['String']>;
  phone?: Maybe<Scalars['String']>;
  continent?: Maybe<Continent>;
  currency?: Maybe<Scalars['String']>;
  languages?: Maybe<Array<Maybe<Language>>>;
  emoji?: Maybe<Scalars['String']>;
  emojiU?: Maybe<Scalars['String']>;
  states?: Maybe<Array<Maybe<State>>>;
};

export type Language = {
   __typename?: 'Language';
  code?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  native?: Maybe<Scalars['String']>;
  rtl?: Maybe<Scalars['Int']>;
};

export type Query = {
   __typename?: 'Query';
  continents?: Maybe<Array<Maybe<Continent>>>;
  continent?: Maybe<Continent>;
  countries?: Maybe<Array<Maybe<Country>>>;
  country?: Maybe<Country>;
  languages?: Maybe<Array<Maybe<Language>>>;
  language?: Maybe<Language>;
};


export type QueryContinentArgs = {
  code?: Maybe<Scalars['String']>;
};


export type QueryCountryArgs = {
  code?: Maybe<Scalars['String']>;
};


export type QueryLanguageArgs = {
  code?: Maybe<Scalars['String']>;
};

export type State = {
   __typename?: 'State';
  code?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
  country?: Maybe<Country>;
};